+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+
+##[1.6.9] - 2020-07-26 WangXH <kurapica125@outlook.com>
+### Changed
+- The `__AutoCache__` attribute can be used on method or functions, so the results will be cached for arguments, the func will return the cached value with the same arguments.
+
+
 ##[1.6.8] - 2020-07-06 WangXH <kurapica125@outlook.com>
 ### Changed
 - Improve the web framework, `__FileView__` added to use view as download file.